Dropbox Daily and monthly backups
Dropbox will have 2 directory’s one for monthly and one for daily
updates each folder has a limit of backups.
If the current date is the first day of the month the backup will upload
to the monthly folder if not it will end up in the daily backup folder.
When the upload has finished it will check to see if the folder has
reached the file limit if so it will delete the oldest file.